refactor: 删除废弃代码
